requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# LLM and AI frameworks
langchain==0.1.4
langchain-community==0.0.16
ollama==0.1.6
pyyaml==6.0.1

# Vector stores and embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.8.0

# Document processing
pypdf==4.0.1
pdfplumber==0.10.3
python-docx==1.1.0
unstructured==0.11.8

# Text processing and NLP
tiktoken==0.5.2
spacy==3.7.2
nltk==3.8.1

# Reranking and search
cohere==4.47
rank-bm25==0.2.2

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
tqdm==4.66.1
tenacity==8.2.3

# Web interface
jinja2==3.1.3
aiofiles==23.2.1
aiohttp==3.9.1
beautifulsoup4>=4.12.0
requests>=2.31.0
